Address PR #148 review comments: optimize readGoodColor, improve type safety, and refactor logic
- Rename readGoodColor.ts to read_good_color.ts to match project conventions
- Optimize readGoodColor to use persistent hidden element instead of creating/removing DOM elements for each call
- Extract parseHexOrRgb and luminance functions to module level to avoid closure issues
- Fix type safety: replace 'any' types with MutableAvailableCity import and proper typing
- Simplify letter good color logic: clarify separation between primaryGood (from map) and letterGood (from availableCities)
- Remove unused goodClass from HeaderHex SVG
- All tests passing
